pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "harness-helm"
version = "0.1.0"
description = "A Textual TUI for observing harness work items."
readme = "README.md"
requires-python = ">=3.11"
keywords = ["harness", "textual", "tui", "markdown", "requirements"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: Other/Proprietary License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: User Interfaces",
  "Topic :: Terminals",
]
dependencies = [
  "PyYAML>=6.0",
  "textual>=0.61.0",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.0",
  "pytest>=8.0",
  "ruff>=0.11.0",
]

[project.scripts]
helm = "helm.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/helm"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
